Reduce precision and improve numerical stability

+++ demo/src/chunk_ray_tracer.rs
@@ -4,21 +4,21 @@ pub trait ChunkRayTracer {
     fn trace_ray_in_chunk(
         &self,
         chunk_data: ChunkData,
-        pos: &na::Vector3<f64>,
-        dir: &na::Vector3<f64>,
+        pos: &na::Vector3<f32>,
+        dir: &na::Vector3<f32>,
         handle: &mut RayTracingResultHandle,
     );
 
-    fn max_radius(&self) -> f64;
+    fn max_radius(&self) -> f32;
 }
 
 pub struct RayTracingResult {
-    pub t: f64,
+    pub t: f32,
     pub intersection: Option<RayTracingIntersection>,
 }
 
 impl RayTracingResult {
-    pub fn new(t: f64) -> Self {
+    pub fn new(t: f32) -> Self {
         RayTracingResult {
             t,
             intersection: None,
@@ -28,24 +28,24 @@ impl RayTracingResult {
 
 pub struct RayTracingIntersection {
     pub voxel_coords: [usize; 2],
-    pub normal: na::Vector3<f64>,
+    pub normal: na::Vector3<f32>,
 }
 
 pub struct RayTracingResultHandle<'a> {
     result: &'a mut RayTracingResult,
-    transform: na::Matrix3<f64>,
+    transform: na::Matrix3<f32>,
 }
 
 impl<'a> RayTracingResultHandle<'a> {
-    pub fn t(&self) -> f64 {
+    pub fn t(&self) -> f32 {
         self.result.t
     }
 
-    pub fn new(result: &'a mut RayTracingResult, transform: na::Matrix3<f64>) -> Self {
+    pub fn new(result: &'a mut RayTracingResult, transform: na::Matrix3<f32>) -> Self {
         RayTracingResultHandle { result, transform }
     }
 
-    pub fn update(&mut self, t: f64, voxel_coords: [usize; 2], normal: na::Vector3<f64>) {
+    pub fn update(&mut self, t: f32, voxel_coords: [usize; 2], normal: na::Vector3<f32>) {
         self.result.t = t;
         self.result.intersection = Some(RayTracingIntersection {
             voxel_coords,
@@ -53,7 +53,7 @@ impl<'a> RayTracingResultHandle<'a> {
         });
     }
 
-    pub fn dependent_handle(&mut self, transform: na::Matrix3<f64>) -> RayTracingResultHandle {
+    pub fn dependent_handle(&mut self, transform: na::Matrix3<f32>) -> RayTracingResultHandle {
         RayTracingResultHandle {
             result: self.result,
             transform: self.transform * transform,
